[Feat] 관리자 회원관리

+++ src/main/java/org/nurim/nurim/repository/AdminRepository.java
@@ -3,5 +3,5 @@ package org.nurim.nurim.repository;
 import org.nurim.nurim.domain.entity.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
 }
